Moved UI (and main) to separate package
