+++ src/main/java/dtu/ui/TerminalGUI.java
@@ -1,6 +1,6 @@
 // Author: Kristian Sofus Knudsen
 
-package dtu.scheduler;
+package dtu.ui;
 
 import java.util.Scanner;
